amman-client: module added and relay client pulled into it
