+++ amman-client/README.md
@@ -1,0 +1,8 @@
+# Amman Client
+
+
+**Client functionality of amman**, interacts with the running validator, relay and storage.
+
+## LICENSE
+
+Apache-2.0
