+++ amman-client/src/relay/api.ts
@@ -1,2 +1,1 @@
 export * from './client'
-export * from './consts'
