+++ amman-client/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "baseUrl": "./",
+    "outDir": "./dist/",
+    "module": "commonjs",
+    "noEmit": false,
+    "declaration": true
+  },
+  "include": ["src"]
+}
