+++ amman-common/src/consts/index.ts
@@ -1,6 +1,2 @@
-// -----------------
-// Mock Storage
-// -----------------
-export const AMMAN_STORAGE_PORT = 50475
-export const AMMAN_STORAGE_URI = `http://localhost:${AMMAN_STORAGE_PORT}`
-export const AMMAN_STORAGE_UPLOAD_URI = `http://localhost:${AMMAN_STORAGE_PORT}/upload`
+export * from './relay'
+export * from './storage'
