+++ amman-common/src/consts/storage.ts
@@ -1,0 +1,3 @@
+export const AMMAN_STORAGE_PORT = 50475
+export const AMMAN_STORAGE_URI = `http://localhost:${AMMAN_STORAGE_PORT}`
+export const AMMAN_STORAGE_UPLOAD_URI = `http://localhost:${AMMAN_STORAGE_PORT}/upload`
