+++ amman-common/src/types/index.ts
@@ -1,1 +1,1 @@
-export const TODO = true
+export * from './relay'
