+++ amman/package.json
@@ -10,6 +10,7 @@
   "scripts": {
     "check:publish-ready": "yarn build",
     "preversion": "yarn check:publish-ready",
+    "postversion": "git add package.json && git commit -m \"chore: update $npm_package_name to v$npm_package_version\" && git tag $npm_package_name@$npm_package_version",
     "prepublishOnly": "yarn check:publish-ready",
     "postpublish": "git push origin && git push origin --tags",
     "build": "rimraf dist && tsc -p tsconfig.json",
@@ -52,6 +53,7 @@
     "@solana/spl-token": "^0.2.0",
     "@solana/spl-token-registry": "^0.2.2405",
     "@solana/web3.js": "^1.35.0",
+    "@metaplex-foundation/amman-client": "^0.0.0",
     "@metaplex-foundation/amman-common": "^0.0.0",
     "ansi-colors": "^4.1.1",
     "bn.js": "^5.2.0",
