+++ amman/src/amman.ts
@@ -3,7 +3,6 @@ export { Change } from './accounts/state'
 export * from './api'
 export * from './asserts'
 export * from './diagnostics'
-export * from './relay/consts'
 export * from './storage/mock-driver'
 export * from './transactions'
 export * from './types'
