+++ amman/src/types.ts
@@ -4,7 +4,10 @@ import { RelayConfig } from './relay/types'
 import { StorageConfig } from './storage'
 import { ValidatorConfig } from './validator/types'
 
-export { RelayAccountState, AccountDiff } from './relay/types'
+export {
+  RelayAccountState,
+  AccountDiff,
+} from '@metaplex-foundation/amman-common'
 
 /**
  * Amman Config
